fix: provide simulate scan operations test case in CursorMap.simulateScanSessions fix: modify evictMapping arguments ts => cursor fix: modify struct CursorMapping.kvstoreId int => size_t fix: provide more debug getter, scoped by #TENDIS_DEBUG

+++ src/tendisplus/commands/scan.cpp
@@ -516,8 +516,7 @@ class ScanCommand : public Command {
     }
 
     if (cursor && !batch.empty()) {
-      cursorMap.addMapping(cursor, static_cast<int>(id),
-                            batch.back().encode(), sess->id());
+      cursorMap.addMapping(cursor, id, batch.back().encode(), sess->id());
       batch.pop_back();
     }
